requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
discovery==0.0.4
Django==3.2.5
django-compat==1.0.15
djangorestframework==3.13.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
notifiers==1.*
oauth2client==4.1.3
protobuf==4.21.5
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.2.1
requests==2.*
rsa==4.9
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
xmltodict==0.13.0
